backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
razorpay==1.4.1
boto3==1.34.0
Pillow==10.1.0
aiofiles==23.2.1
emails==0.6
twilio==8.10.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2